+++ lib/proto/http.h
@@ -26,8 +26,8 @@ struct http_pack_ctx *_http_parsehead(buffer_ctx *buf, int32_t *status, int32_t 
 int32_t _http_check_keyval(http_header_ctx *head, const char *key, const char *val);
 
 buf_ctx *http_status(struct http_pack_ctx *pack);
-size_t http_nheader(struct http_pack_ctx *pack);
-http_header_ctx *http_header_at(struct http_pack_ctx *pack, size_t pos);
+uint32_t http_nheader(struct http_pack_ctx *pack);
+http_header_ctx *http_header_at(struct http_pack_ctx *pack, uint32_t pos);
 char *http_header(struct http_pack_ctx *pack, const char *header, size_t *lens);
 int32_t http_chunked(struct http_pack_ctx *pack);
 void *http_data(struct http_pack_ctx *pack, size_t *lens);
